Various web changes as myelin.co.nz is down and stylesheets aren't being loaded (!)
Added Scott Lewis's changepasswd.py; now we can change passwords (while the server is stopped)

pycs.net index removed; it didn't need to be in there.  Makefile changed to copy local_index.html to the main server index if it exists, otherwise index.html -- that way people can have a local version and still keep up-to-date with the official one.

README now has a picture of an owl, courtesy of Kyoko :)